migrate cache injector to a cache wrapper
