+++ lib/stitches/railtie.rb
@@ -1,12 +1,11 @@
 require 'stitches/api_key'
 require 'stitches/valid_mime_type'
-require 'stitches/cache_injector'
+require 'stitches/api_key_cache_wrapper'
 
 module Stitches
   class Railtie < Rails::Railtie
     config.app_middleware.use Stitches::ApiKey
     config.app_middleware.use Stitches::ValidMimeType
 
-    Stitches::CacheInjector.inject if defined? ApiClient
   end
 end
