+++ spec/api_key_spec.rb
@@ -11,8 +11,6 @@ unless defined? ApiClient
       ["enabled"]
     end
   end
-
-  Stitches::CacheInjector.inject
 end
 
 describe Stitches::ApiKey do
